# docs/manhes/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Manga ingestion and catalog API.",
        "title": "Manhes API",
        "contact": {},
        "version": "2.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/dictionary": {
            "get": {
                "description": "Results are permanently stored in the dictionary. Re-searching a known",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dictionary"
                ],
                "summary": "Search all scrapers and auto-upsert results into dictionary",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search query",
                        "name": "q",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.DictionaryResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/dictionary/{dictionaryId}": {
            "post": {
                "description": "Re-fetches source stats from all known scrapers and searches",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dictionary"
                ],
                "summary": "Refresh a dictionary entry",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Dictionary entry ID",
                        "name": "dictionaryId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.DictionaryResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httputil.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/manga": {
            "get": {
                "description": "Returns a paginated list of manga. Supports filtering by id, q, genre, author, state and sorting.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "manga"
                ],
                "summary": "List manga",
                "parameters": [
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "Filter by dictionary_id (comma-separated or repeated)",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search by title or description",
                        "name": "q",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "Filter by genre (comma-separated or repeated)",
                        "name": "genre",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "Filter by author (comma-separated or repeated)",
                        "name": "author",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "Filter by state (unavailable|fetching|available)",
                        "name": "state",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort field: title | updatedAt | createdAt (default: title)",
                        "name": "sortBy",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort order: asc | desc (default: asc)",
                        "name": "sortOrder",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number (default 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page (default 20, max 100)",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.MangaListResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/manga/{mangaId}": {
            "get": {
                "description": "mangaId is the dictionary entry UUID. Unavailable/fetching manga return partial detail.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "manga"
                ],
                "summary": "Get manga detail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Dictionary entry UUID",
                        "name": "mangaId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.MangaDetailResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httputil.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/manga/{mangaId}/{lang}": {
            "get": {
                "description": "Returns all uploaded chapters for a specific language. Requires manga to be in available state.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "manga"
                ],
                "summary": "List uploaded chapters for a language",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Dictionary entry UUID",
                        "name": "mangaId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Language code (e.g. en, fr)",
                        "name": "lang",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.ChapterListResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httputil.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/manga/{mangaId}/{lang}/read": {
            "get": {
                "description": "Returns page URLs for the requested chapter along with prev/next navigation links. Requires manga to be in available state.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "manga"
                ],
                "summary": "Get chapter pages",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Dictionary entry UUID",
                        "name": "mangaId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Language code (e.g. en, fr)",
                        "name": "lang",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "Chapter number",
                        "name": "chapter",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.ChapterReadResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httputil.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httputil.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httputil.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.ChapterItem": {
            "type": "object",
            "properties": {
                "chapter": {
                    "type": "string"
                },
                "page_count": {
                    "type": "integer"
                },
                "uploaded_at": {
                    "type": "string"
                }
            }
        },
        "domain.ChapterListResponse": {
            "type": "object",
            "properties": {
                "chapters": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.ChapterItem"
                    }
                },
                "id": {
                    "type": "string"
                },
                "lang": {
                    "type": "string"
                }
            }
        },
        "domain.ChapterReadResponse": {
            "type": "object",
            "properties": {
                "chapter": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "lang": {
                    "type": "string"
                },
                "next_chapter": {
                    "type": "string"
                },
                "pages": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "prev_chapter": {
                    "type": "string"
                }
            }
        },
        "domain.ChapterStats": {
            "type": "object",
            "properties": {
                "available": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "domain.DictionaryResponse": {
            "type": "object",
            "properties": {
                "best_source": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "chapters_by_lang": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                },
                "cover_url": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "source_stats": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/domain.SourceStat"
                    }
                },
                "sources": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "domain.MangaDetailResponse": {
            "type": "object",
            "properties": {
                "authors": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "cover_url": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "genres": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "string"
                },
                "languages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.MangaLangResponse"
                    }
                },
                "state": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "domain.MangaLangResponse": {
            "type": "object",
            "properties": {
                "available_chapters": {
                    "type": "integer"
                },
                "lang": {
                    "type": "string"
                },
                "latest_update": {
                    "type": "string"
                },
                "total_chapters": {
                    "type": "integer"
                },
                "uploaded_chapters": {
                    "type": "integer"
                }
            }
        },
        "domain.MangaListResponse": {
            "type": "object",
            "properties": {
                "itemCount": {
                    "type": "integer"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.MangaSummary"
                    }
                },
                "pageNumber": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "pageTotal": {
                    "type": "integer"
                }
            }
        },
        "domain.MangaSummary": {
            "type": "object",
            "properties": {
                "authors": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "chapters_by_lang": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/domain.ChapterStats"
                    }
                },
                "cover_url": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "dictionary_id": {
                    "type": "string"
                },
                "genres": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "domain.SourceStat": {
            "type": "object",
            "properties": {
                "chapters_by_lang": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                },
                "err": {
                    "type": "string"
                },
                "fetched_at": {
                    "type": "string"
                }
            }
        },
        "httputil.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}